Remove legacy maven tests
