--- Tests-Legacy/L0/MavenV3/data/analysis_details.json
@@ -1,1 +1,0 @@
-{"projectStatus":{"status":"WARN","conditions":[{"status":"WARN","metricKey":"lines","comparator":"GT","warningThreshold":"1","actualValue":"71"},{"status":"WARN","metricKey":"sqale_index","comparator":"GT","warningThreshold":"0","actualValue":"87"}],"periods":[{"index":1,"mode":"previous_version","date":"2016-07-07T16:34:23+0100"},{"index":2,"mode":"previous_analysis","date":"2016-08-18T17:24:34+0100","parameter":"2016-08-18"},{"index":3,"mode":"days","date":"2016-07-07T16:36:20+0100","parameter":"30"}]}}
