--- Tests-Legacy/L0/MavenV3/data/task_details.json
@@ -1,1 +1,0 @@
-{"task":{"id":"asdfghjklqwertyuiopz","type":"REPORT","componentId":"AVVKNKgxirtVJMydykhN","componentKey":"test:test","componentName":"test","componentQualifier":"TRK","analysisId":"12345","status":"SUCCESS","submittedAt":"2016-06-29T13:58:35+0000","submitterLogin":"admin","startedAt":"2016-06-29T13:58:37+0000","executedAt":"2016-06-29T13:58:38+0000","executionTimeMs":1471,"logs":false}}
