--- Tests-Legacy/L0/MavenV3/responseCodeCoverage.json
@@ -1,120 +1,0 @@
-{
-    "which": {
-        "mvn": "/home/bin/maven/bin/mvn"
-    },
-    "exec": {
-        "/home/bin/maven/bin/mvn -version": {
-            "code": 0,
-            "stdout": "Maven version 1.0.0"
-        },
-        "/home/bin/maven/bin/mvn -f pom.xml package": {
-            "code": 0,
-            "stdout": "Maven package done"
-        },
-        "/home/bin/maven/bin/mvn -f pom.xml clean package": {
-            "code": 0,
-            "stdout": "Maven package done"
-        },
-        "/home/bin/maven2/bin/mvn -version": {
-            "code": 0,
-            "stdout": "Maven version 1.0.0"
-        },
-        "/home/bin/maven2/bin/mvn -f pom.xml package": {
-            "code": 0,
-            "stdout": "Maven package done"
-        },
-        "/home/bin/maven2/bin/mvn -f pom.xml clean package": {
-            "code": 0,
-            "stdout": "Maven package done"
-        },
-        "/home/bin/maven/bin/mvn -f pom.xml jacoco:report": {
-            "code": 0,
-            "stdout": "Jacoco report done"
-        },
-        "/home/bin/maven/bin/mvn -f pom.xml /o /p /t:i o /n /s package": {
-            "code": 0,
-            "stdout": "Maven package done"
-        },
-        "/home/bin/maven/bin/mvn -f pom.xml /o /p /t /i /o /n /s package": {
-            "code": 0,
-            "stdout": "Maven package done"
-        },
-        "/home/bin/maven/bin/mvn -f pom.xml /o /p /t /i /o /n /s build test package": {
-            "code": 0,
-            "stdout": "Maven package done"
-        },
-        "/home/bin/maven/bin/mvn -f pom.xml FAIL package": {
-            "code": 2,
-            "stdout": "Running Maven...",
-            "stderr": "FAILED"
-        },
-        "/home/bin/maven/bin/mvn -f pom.xml clean package -Dsonar.host.url=http://sonarqubeserver:9000 -Dsonar.login=uname -Dsonar.password=pword sonar:sonar": {
-            "code": 0,
-            "stdout": "Maven package and SQ analysis done"
-        },
-        "/home/bin/maven/bin/mvn -f pom.xml clean package -Dsonar.host.url=http://sonarqubeserver:9000 -Dsonar.login=uname -Dsonar.password=pword -Dsonar.jdbc.url=dbURL -Dsonar.jdbc.username=dbUser -Dsonar.jdbc.password=dbPass sonar:sonar": {
-            "code": 0,
-            "stdout": "Maven package and SQ analysis done"
-        },
-        "/home/bin/maven/bin/mvn -f pom.xml clean package -Dsonar.host.url=http://sonarqubeserver:9000 -Dsonar.login=uname -Dsonar.password=pword -Dsonar.jacoco.reportPath=CCReport43F6D5EF/jacoco.exec sonar:sonar": {
-            "code": 0,
-            "stdout": "Maven package and SQ analysis done"
-        }
-    },
-    "checkPath": {
-        "/home/bin/maven/bin/mvn": true,
-        "/home/bin/maven": true,
-        "/home/bin/maven2": true,
-        "pom.xml": true
-    },
-    "getVariable": {
-        "JAVA_HOME_8_X86": "/user/local/bin/Java8",
-        "agent.buildDirectory": "/user/build",
-        "ENDPOINT_URL_ID1": "http://sonarqubeserver:9000",
-        "ENDPOINT_AUTH_ID1": "{\"scheme\":\"UsernamePassword\", \"parameters\": {\"username\": \"uname\", \"password\": \"pword\"}}",
-        "build.sourcesDirectory": "/user/build/s",
-        "build.artifactStagingDirectory": "/user/build/a",
-        "Agent.TempDirectory": "/tmp",
-        "System.DefaultWorkingDirectory": "/user/build/s"
-    },
-    "find": {
-        "/user/build": [
-            "/user/build/fun/test-123.xml"
-        ]
-    },
-    "match": {
-        "**/TEST-*.xml": [
-            "/user/build/fun/test-123.xml"
-        ]
-    },
-    "exist": {
-        "CCReport43F6D5EF/jacoco.xml": true,
-        "target/site/cobertura/coverage.xml": true
-    },
-    "stats": {
-        "CCReport43F6D5EF/jacoco.xml": {
-            "isFile": true
-        },
-        "target/site/cobertura/coverage.xml": {
-            "isFile": true
-        }
-    },
-    "rmRF": {
-        "target": {
-            "success": true,
-            "message": "success"
-        },
-        "CCReport43F6D5EF": {
-            "success": true,
-            "message": "success"
-        },
-        "target/site/cobertura": {
-            "success": true,
-            "message": "success"
-        },
-        "CCReportPomA4D283EG.xml": {
-            "success": true,
-            "message": "success"
-        }
-    }
-}
