--- Tests-Legacy/L0/MavenV3/responseM2_HOME.json
@@ -1,81 +1,0 @@
-{
-    "exec": {
-        "/home/bin/maven/bin/mvn -version": {
-            "code": 0,
-            "stdout": "Maven version 1.0.0"
-        },
-        "/home/bin/maven/bin/mvn -f pom.xml package": {
-            "code": 0,
-            "stdout": "Maven package done"
-        },
-        "/home/bin/maven2/bin/mvn -version": {
-            "code": 0,
-            "stdout": "Maven version 1.0.0"
-        },
-        "/home/bin/maven/bin/mvn -f pom.xml help:effective-pom": {
-            "code": 0,
-            "stdout": "Effective POMs, after inheritance, interpolation, and profiles are applied:\r\n\r\n<!-- ====================================================================== -->\r\n<!--                                                                        -->\r\n<!-- Generated by Maven Help Plugin on 2017-06-29T09:43:41                  -->\r\n<!-- See: http:\/\/maven.apache.org\/plugins\/maven-help-plugin\/                -->\r\n<!--                                                                        -->\r\n<!-- ====================================================================== -->\r\n\r\n<!-- ====================================================================== -->\r\n<!--                                                                        -->\r\n<!-- Effective POM for project                                              -->\r\n<!-- 'com.microsoft.xplatalm:xplatalmApp:jar:1.0-SNAPSHOT'                  -->\r\n<!--                                                                        -->\r\n<!-- ====================================================================== -->\r\n\r\n<project xmlns=\"http:\/\/maven.apache.org\/POM\/4.0.0\" xmlns:xsi=\"http:\/\/www.w3.org\/2001\/XMLSchema-instance\" xsi:schemaLocation=\"http:\/\/maven.apache.org\/POM\/4.0.0 http:\/\/maven.apache.org\/xsd\/maven-4.0.0.xsd\">\r\n<\/project>\r\n\r\n[INFO] ------------------------------------------------------------------------\r\n[INFO] BUILD SUCCESS\r\n[INFO] ------------------------------------------------------------------------\r\n[INFO] Total time: 0.927 s\r\n[INFO] Finished at: 2017-06-29T09:43:41-04:00\r\n[INFO] Final Memory: 7M\/18M\r\n[INFO] ------------------------------------------------------------------------"
-        },
-        "/home/bin/maven2/bin/mvn -f pom.xml package": {
-            "code": 0,
-            "stdout": "Maven package done"
-        },
-        "/home/bin/maven/bin/mvn -f pom.xml /o /p /t:i o /n /s package": {
-            "code": 0,
-            "stdout": "Maven package done"
-        },
-        "/home/bin/maven/bin/mvn -f pom.xml /o /p /t /i /o /n /s package": {
-            "code": 0,
-            "stdout": "Maven package done"
-        },        
-        "/home/bin/maven/bin/mvn -f pom.xml /o /p /t /i /o /n /s build test package": {
-            "code": 0,
-            "stdout": "Maven package done"
-        },        
-        "/home/bin/maven/bin/mvn -f pom.xml FAIL package": {
-            "code": 2,
-            "stdout": "Running Maven...",
-            "stderr": "FAILED"
-        },
-        "/home/bin/maven/bin/mvn -f pom.xml package -Dsonar.host.url=http://sonarqube/end/point -Dsonar.login=uname -Dsonar.password=pword sonar:sonar":{
-            "code": 0,
-            "stdout": "Maven package and SQ analysis done"
-        },
-        "/home/bin/maven/bin/mvn -f pom.xml package -Dsonar.host.url=http://sonarqube/end/point -Dsonar.login=uname -Dsonar.password=pword -Dsonar.jdbc.url=dbURL -Dsonar.jdbc.username=dbUser -Dsonar.jdbc.password=dbPass sonar:sonar":{
-            "code": 0,
-            "stdout": "Maven package and SQ analysis done"
-        }
-    },
-    "checkPath" : {
-        "/home/bin/maven": true,
-        "/home/bin/maven/bin/mvn": true,
-        "/home/bin/maven2": true,
-        "pom.xml": true
-    },
-    "getVariable" : {
-        "M2_HOME": "/home/bin/maven",
-        "JAVA_HOME_8_X86": "/user/local/bin/Java8",
-        "agent.buildDirectory": "/user/build",
-        "Agent.TempDirectory": "/tmp",
-        "ENDPOINT_URL_ID1": "http://sonarqube/end/point",
-        "ENDPOINT_AUTH_ID1": "{\"scheme\":\"UsernamePassword\", \"parameters\": {\"username\": \"uname\", \"password\": \"pword\"}}",
-        "ENDPOINT_AUTH_SYSTEMVSSCONNECTION": "{\"parameters\":{\"AccessToken\":\"token\"},\"scheme\":\"OAuth\"}",
-        "build.sourcesDirectory": "/user/build/s",
-        "build.artifactStagingDirectory": "/user/build/a",
-        "build.buildNumber": "1",
-        "System.DefaultWorkingDirectory": "/user/build/s",
-        "System.TeamFoundationCollectionUri": "https://xplatalm.visualstudio.com/"
-
-    },
-    "find": {
-        "/user/build": [
-            "/user/build/fun/test-123.xml",
-            "/user/build/a/.mavenInfo"
-        ]
-    },
-    "match": {
-        "**/TEST-*.xml": [
-            "/user/build/fun/test-123.xml"
-        ]
-    }
-}
